+++ src/core/include/cesium/omniverse/FabricGeometryDefinition.h
@@ -1,6 +1,9 @@
 #pragma once
 
+#include "cesium/omniverse/GltfUtil.h"
+
 #include <cstdint>
+#include <set>
 
 namespace CesiumGltf {
 struct MeshPrimitive;
@@ -18,16 +21,18 @@ class FabricGeometryDefinition {
 
     [[nodiscard]] bool hasNormals() const;
     [[nodiscard]] bool hasVertexColors() const;
-    [[nodiscard]] bool getDoubleSided() const;
     [[nodiscard]] uint64_t getTexcoordSetCount() const;
+    [[nodiscard]] const std::set<VertexAttributeInfo>& getCustomVertexAttributes() const;
 
     bool operator==(const FabricGeometryDefinition& other) const;
 
   private:
     bool _hasNormals{false};
     bool _hasVertexColors{false};
-    bool _doubleSided{false};
     uint64_t _texcoordSetCount{0};
+
+    // std::set is sorted which is important for checking FabricGeometryDefinition equality
+    std::set<VertexAttributeInfo> _customVertexAttributes;
 };
 
 } // namespace cesium::omniverse