+++ src/core/src/FabricGeometryDefinition.cpp
@@ -14,14 +14,13 @@ namespace cesium::omniverse {
 FabricGeometryDefinition::FabricGeometryDefinition(
     const CesiumGltf::Model& model,
     const CesiumGltf::MeshPrimitive& primitive,
-    bool smoothNormals) {
-    const auto materialInfo = GltfUtil::getMaterialInfo(model, primitive);
-    _hasNormals = GltfUtil::hasNormals(model, primitive, smoothNormals);
-    _hasVertexColors = GltfUtil::hasVertexColors(model, primitive, 0);
-    _doubleSided = materialInfo.doubleSided;
-    _texcoordSetCount = GltfUtil::getTexcoordSetIndexes(model, primitive).size() +
-                        GltfUtil::getImageryTexcoordSetIndexes(model, primitive).size();
-}
+    bool smoothNormals)
+    : _hasNormals(GltfUtil::hasNormals(model, primitive, smoothNormals))
+    , _hasVertexColors(GltfUtil::hasVertexColors(model, primitive, 0))
+    , _texcoordSetCount(
+          GltfUtil::getTexcoordSetIndexes(model, primitive).size() +
+          GltfUtil::getImageryTexcoordSetIndexes(model, primitive).size())
+    , _customVertexAttributes(GltfUtil::getCustomVertexAttributes(model, primitive)) {}
 
 bool FabricGeometryDefinition::hasNormals() const {
     return _hasNormals;
@@ -31,32 +30,17 @@ bool FabricGeometryDefinition::hasVertexColors() const {
     return _hasVertexColors;
 }
 
-bool FabricGeometryDefinition::getDoubleSided() const {
-    return _doubleSided;
-}
-
-[[nodiscard]] uint64_t FabricGeometryDefinition::getTexcoordSetCount() const {
+uint64_t FabricGeometryDefinition::getTexcoordSetCount() const {
     return _texcoordSetCount;
 }
 
+const std::set<VertexAttributeInfo>& FabricGeometryDefinition::getCustomVertexAttributes() const {
+    return _customVertexAttributes;
+}
+
 bool FabricGeometryDefinition::operator==(const FabricGeometryDefinition& other) const {
-    if (_hasNormals != other._hasNormals) {
-        return false;
-    }
-
-    if (_hasVertexColors != other._hasVertexColors) {
-        return false;
-    }
-
-    if (_doubleSided != other._doubleSided) {
-        return false;
-    }
-
-    if (_texcoordSetCount != other._texcoordSetCount) {
-        return false;
-    }
-
-    return true;
+    return _hasNormals == other._hasNormals && _hasVertexColors == other._hasVertexColors &&
+           _texcoordSetCount == other._texcoordSetCount && _customVertexAttributes == other._customVertexAttributes;
 }
 
 } // namespace cesium::omniverse